fix: adding api to add item after having cart id
